Move labeling model imports to snorkel/labeling/model
