+++ snorkel/labeling/__init__.py
@@ -4,10 +4,4 @@ from .analysis import LFAnalysis  # noqa: F401
 from .apply.core import LFApplier  # noqa: F401
 from .apply.pandas import PandasLFApplier  # noqa: F401
 from .lf.core import LabelingFunction, labeling_function  # noqa: F401
-from .model.baselines import (  # noqa: F401
-    MajorityClassVoter,
-    MajorityLabelVoter,
-    RandomVoter,
-)
-from .model.label_model import LabelModel  # noqa: F401
 from .utils import filter_unlabeled_dataframe  # noqa: F401
