+++ snorkel/labeling/model/__init__.py
@@ -1,0 +1,2 @@
+from .baselines import MajorityClassVoter, MajorityLabelVoter, RandomVoter  # noqa: F401
+from .label_model import LabelModel  # noqa: F401
