+++ test/labeling/model/test_baseline.py
@@ -2,7 +2,7 @@ import unittest
 
 import numpy as np
 
-from snorkel.labeling import MajorityClassVoter, MajorityLabelVoter, RandomVoter
+from snorkel.labeling.model import MajorityClassVoter, MajorityLabelVoter, RandomVoter
 
 
 class BaselineModelTest(unittest.TestCase):
