+++ test/labeling/model/test_label_model.py
@@ -9,7 +9,7 @@ import torch
 import torch.nn as nn
 import torch.optim as optim
 
-from snorkel.labeling import LabelModel
+from snorkel.labeling.model import LabelModel
 from snorkel.labeling.model.label_model import TrainConfig
 from snorkel.synthetic.synthetic_data import generate_simple_label_matrix
